src/main/java/model/domain/user/Person.java
package model.domain.user;

/**
 * Person.java
 * A class that represents a person
 * Created by Michael on 2015/08/07.
 */
public class Person
{
    private String email;
    private String firstName;
    private String lastName;

    public Person(String firstName, String lastName, String email)
    {
        this.firstName = firstName;
        this.lastName = lastName;
        this.email = email;
    }

    public void setEmail(String email)
    {
        this.email = email;
    }

    public String getEmail()
    {
        return email;
    }

    public void setFirstName(String firstName)
    {
        this.firstName = firstName;
    }

    public String getFirstName()
    {
        return firstName;
    }

    public void setLastName(String lastName)
    {
        this.lastName = lastName;
    }

    public String getLastName()
    {
        return lastName;
    }
}
